[android] Fix reading 64-bit integers from offline db

+++ app-android/tutashared/src/main/java/de/tutao/tutashared/offline/TaggedSqlValue.kt
@@ -16,7 +16,7 @@ sealed class TaggedSqlValue {
 
 	@Serializable
 	@SerialName("SqlNum")
-	data class Num(val value: Int) : TaggedSqlValue()
+	data class Num(val value: Long) : TaggedSqlValue()
 
 	@Serializable
 	@SerialName("SqlBytes")
